Final version

+++ UnitConverter/app/src/main/java/com/orlik/converter/viewModels/VolumeViewModel.kt
@@ -1,10 +1,11 @@
 package com.orlik.converter.viewModels
 
 import androidx.lifecycle.ViewModel
+import java.math.BigDecimal
 
 class VolumeViewModel : ViewModel() {
-    var editValue: Float = 0.0f
-    var resultValue: Float = 0.0f
+    var editValue: BigDecimal = BigDecimal(0)
+    var resultValue: BigDecimal = BigDecimal(0)
     var convertFrom: Int = 2
     var convertTo: Int = 2
 }
